requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib
numpy
opencv-python==3.4.4.19
opencv-contrib-python==3.4.4.19
Pillow
PyYAML
scipy
torch
torchvision
tqdm
pafy>=0.5.5

# extras --------------------------------------
thop  # FLOPS computation
pycocotools  # COCO mAP